Changes in Footer: rename file name and add the sponsor's logos.
