+++ src/components/NotFound.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Header } from "./Header/Header";
-import Footer from "./Footer";
+import Footer from "./Footer/Footer";
 
 export default function NotFound() {
     return (
