+++ src/components/Programma/Programma.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Footer from "../Footer";
+import Footer from "../Footer/Footer";
 import {Header} from "../Header/Header";
 import {useTranslation} from 'react-i18next';
